Refactor home with single material app (fix #631, fix #625)
